Remove instance count
This (along with the previous instance var) introduced more
complexity than seemed necessary. Saying that the nth and (n+1)th
iteration of an action may not yield the same result is confusing
and adds more stuff to account for when testing.

The original purpose of tracking an instance var was to help loops
recognize when they were being called multiple times within the
same iteration -- e.g. A returns continuation actions B1 and B2,
and both B1 and B2 return a continuation action of C. In this case,
presumably, it might not performant or correct to run C twice.

However, those rationales could also apply if B1 and B2 returned
effects to dispatch asynchronously, in which case instance
information wouldn't be available anyway. In that case, C has to
rely on prior state to avoid unnecessary or incorrect state changes
anyway. Given that, we're better off just having C always look to
state than an instance var. It's also closer to idiomatic Redux.